+++ functions/Export-MySQLiteDB.ps1
@@ -24,8 +24,11 @@ Function Export-MySQLiteDB {
     Begin {
         Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
         Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Opening a connection to $Path"
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Detected culture $(Get-Culture)"
+        if ($MyInvocation.CommandOrigin -eq 'Runspace') {
+            #Hide this metadata when the command is called from another command
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Detected culture $(Get-Culture)"
+        }
         Try {
             #always open the database
             $conn = Open-MySQLiteDB -Path $path -ErrorAction Stop -WhatIf:$False
@@ -39,15 +42,27 @@ Function Export-MySQLiteDB {
         if ($conn.state -eq 'open') {
             #initialize a hashtable
             $hash = @{}
+            $tableInfo = @()
             Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Getting tables"
             $tables = (Get-MySQLiteTable -Connection $conn -KeepAlive -ErrorAction Stop).Name
             Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Exporting $Path"
             foreach ($table in $tables) {
+                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Getting table schema"
+                $schema =  Invoke-MySQLiteQuery -Query "pragma table_Info($table)" -Connection $conn -KeepAlive  |
+                Select-Object -Property Name,Type,notnull,dflt_value,pk
                 $query = "Select * from $table"
                 Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $query"
                 Try {
                     $data = Invoke-MySQLiteQuery -Query $query -Connection $conn -KeepAlive -ErrorAction Stop
-                    $hash.add($table, $data)
+                    $details = [PSCustomObject]@{
+                        Name = $table
+                        Schema = $schema
+                        Data = $data
+                    }
+                    $tableInfo += $details
+                    #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Adding $table to hashtable"
+                    #$hash.add($table, $details)
+                   # $hash.add($table, $data)
                 }
                 Catch {
                     Write-Warning "There was an error invoking the last query."
@@ -55,9 +70,11 @@ Function Export-MySQLiteDB {
                     Throw $_
                 }
             }
+
+            $hash.add('Tables', $tableInfo)
             Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Saving to $Destination"
             if ($PSCmdlet.ShouldProcess($Destination, "Export database $path")) {
-                $hash | ConvertTo-Json -Depth 100 | Set-Content -Path $destination -Encoding utf8
+                [PSCustomObject]$hash | ConvertTo-Json -Depth 100 | Set-Content -Path $destination -Encoding utf8
                 if ($PassThru) {
                     Get-Item -Path $Destination
                 }