+++ functions/Invoke-MySQLiteQuery.ps1
@@ -36,30 +36,38 @@ Function Invoke-MySQLiteQuery {
         [string]$As = "object"
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($MyInvocation.MyCommand)"
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Running under PowerShell version $($PSVersionTable.PSVersion)"
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Detected culture $(Get-Culture)"
+        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
+        if ($MyInvocation.CommandOrigin -eq 'Runspace') {
+            #Hide this metadata when the command is called from another command
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Detected culture $(Get-Culture)"
+        }
         $exceptionDelegate = {
             param([System.Management.Automation.ErrorRecord]$errRecord)
             # if inner exception is System.Data.SQLite.SQLiteException (0x800007BF): SQL logic, help user by telling them what the error is
             if ($errRecord.Exception.InnerException -is [System.Data.SQLite.SQLiteException]) {
                 $errTxt = $errRecord.Exception.InnerException.Message -split "`n"
-                Write-Warning $errTxt[0]
+                Write-Error $errTxt[0]
                 $syntaxErr = $errTxt[1] | Select-String -Pattern '(?<=")[^"]+(?=")'
-                $syntaxErr = $syntaxErr.Matches.Value
-                # highlight offending token
-                $query -replace ([regex]::Escape($syntaxErr)), "`e[7m`$0`e[0;93m" | Write-Warning
-                # generate a pointer caret to the offending token
-                ' ' * $query.IndexOf($syntaxErr) + '^' | Write-Warning
+                if ($syntaxErr) {
+                    $syntaxErr = $syntaxErr.Matches.Value
+                    # highlight offending token
+                    $query -replace ([regex]::Escape($syntaxErr)), "`e[7m`$0`e[0;93m" | Write-Error
+                    # generate a pointer caret to the offending token
+                    ' ' * $query.IndexOf($syntaxErr) + '^' | Write-Error
+                }
+                else {
+                    Write-Error $errTxt[1]
+                }
             }
             else {
-                Write-Warning $_.Exception.Message
+                Write-Error $_.Exception.Message
             }
         }
     } #begin
     Process {
         if ($PSCmdlet.ParameterSetName -eq 'file') {
-            Write-Verbose "[$((Get-Date).TimeOfDay)] Using file $path"
+            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Using file $path"
             $file = resolvedb -path $path
 
             If ($file.exists) {
@@ -70,11 +78,13 @@ Function Invoke-MySQLiteQuery {
             }
         }
         else {
-            Write-Verbose "[$((Get-Date).TimeOfDay)] Using connection $($connection.ConnectionString)"
-            Write-Verbose "[$((Get-Date).TimeOfDay)] KeepAlive is $KeepAlive"
+            if ($MyInvocation.CommandOrigin -eq 'Runspace') {
+                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Using connection $($connection.ConnectionString)"
+                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] KeepAlive is $KeepAlive"
+            }
         }
 
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Invoke query '$query'"
+        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Invoke query '$query'"
         if ($connection.state -eq 'Open') {
 
             $cmd = $connection.CreateCommand()
@@ -85,24 +95,26 @@ Function Invoke-MySQLiteQuery {
                 "^([Ss]elect (\w+|\*)|(@@\w+ AS))|([Pp]ragma \w+)" {
                     #   "^Select (\w+|\*)|(@@\w+ AS)" {
                     if ($As -eq "datatable") {
-                        Write-Verbose "[$((Get-Date).TimeOfDay)] Datatable output"
+                        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Datatable output"
                         $ds = New-Object System.Data.DataSet
                         $da = New-Object System.Data.SQLite.SQLiteDataAdapter($cmd)
                         [void]$da.fill($ds)
                         $ds.Tables
                     }
                     else {
-                        Write-Verbose "[$((Get-Date).TimeOfDay)] ExecuteReader"
+                        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] ExecuteReader"
                         try {
                             $reader = $cmd.ExecuteReader()
                         }
                         catch [System.Management.Automation.MethodInvocationException] {
-                            $exceptionDelegate.Invoke($_)
+                            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Caught MethodInvocationException"
+                            $global:e = $_
+                            #$exceptionDelegate.Invoke($_)
+                            & $exceptionDelegate $_
                             return
                         }
                         #convert datarows to a custom object
                         while ($reader.read()) {
-
                             $h = [ordered]@{}
                             for ($i = 0; $i -lt $reader.FieldCount; $i++) {
                                 $col = $reader.GetName($i)
@@ -123,13 +135,13 @@ Function Invoke-MySQLiteQuery {
                     Break
                 }
                 "@@" {
-                    Write-Verbose "[$((Get-Date).TimeOfDay)] ExecuteScalar"
+                    Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] ExecuteScalar"
                     [void]$cmd.ExecuteScalar()
                     Break
                 }
                 Default {
                     if ($PSCmdlet.ShouldProcess($query)) {
-                        Write-Verbose "[$((Get-Date).TimeOfDay)] ExecuteNonQuery"
+                        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] ExecuteNonQuery"
                         #modify query to use Transactions
                         $Revised = "BEGIN TRANSACTION;$($cmd.CommandText);COMMIT;"
                         Write-Verbose "[$((Get-Date).TimeOfDay)] $Revised"
@@ -139,7 +151,8 @@ Function Invoke-MySQLiteQuery {
                             [void]$cmd.ExecuteNonQuery()
                         }
                         catch [System.Management.Automation.MethodInvocationException] {
-                            $exceptionDelegate.Invoke($_)
+                            #$exceptionDelegate.Invoke($_)
+                            & $exceptionDelegate $_
                         }
                         catch {
                             Write-Warning $_.Exception.message
@@ -152,9 +165,9 @@ Function Invoke-MySQLiteQuery {
     End {
         #if the connection was passed as a parameter, do not close it. The generating command is responsible for managing the connection.
         if ( (($connection.state -eq 'Open') -AND ($PSCmdlet.ParameterSetName -eq 'file')) -OR (($connection.state -eq 'Open') -AND (-Not $KeepAlive)) ) {
-            Write-Verbose "[$((Get-Date).TimeOfDay)] Closing database connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Closing database connection"
             closedb -connection $connection -cmd $cmd
         }
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($MyInvocation.MyCommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
     } #end
 }