+++ functions/Open-MySQLiteDB.ps1
@@ -10,22 +10,25 @@ Function Open-MySQLiteDB {
         [string]$Path
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] $($MyInvocation.MyCommand)"
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Running under PowerShell version $($PSVersionTable.PSVersion)"
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Detected culture $(Get-Culture)"
+        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand)"
+        if ($MyInvocation.CommandOrigin -eq 'Runspace') {
+            #Hide this metadata when the command is called from another command
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Detected culture $(Get-Culture)"
+        }
     } #begin
 
     Process {
         $db = resolvedb $Path
         if ($db.exists) {
             if ($PSCmdlet.ShouldProcess($db.path)) {
-                Write-Verbose "[$((Get-Date).TimeOfDay)] Opening $Path"
+                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Opening $Path"
                 opendb $db.path
             }
         }
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($MyInvocation.MyCommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
     } #end
 }